api/requirements.txt
djangorestframework
pillow
python-decouple